+++ src/control_plane/project_infrastructure/mod.rs
@@ -42,5 +42,7 @@ pub(crate) use project_service_provisioning_job::ProjectServiceProvisioningJob;
 mod live_engine_tests;
 mod materialize_project_service_configurations;
 #[cfg(test)]
+mod search_live_engine_tests;
+#[cfg(test)]
 mod tests;
 pub(crate) use materialize_project_service_configurations::materialize_project_service_configurations;